requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
coverage==7.10.5
Django==5.2.2
django-appconf==1.1.0
django-cors-headers==4.7.0
django-ipware==7.0.1
django-redis==5.4.0
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-recaptcha==4.0.2
gunicorn==23.0.0
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.0
python-ipware==3.0.0
pytz==2025.2
redis==6.2.0
rq==2.3.3
sqlparse==0.5.3
whitenoise==6.9.0
